requirements.txt
FPDF
PdfReader
streamlit
google-generativeai
python-dotenv
langchain
faiss-cpu
langchain_google_genai
python-box